[refac] 리소스 출력 메서드를  util 패키지로  분리
